Remove dependency of v05 on v04 (#171)
