+++ src/ome_zarr_models/_v05/well.py
@@ -1,15 +1,17 @@
 from pydantic_zarr.v2 import ArraySpec, GroupSpec
 
-import ome_zarr_models.v04.well
 from ome_zarr_models._v05.base import BaseGroupv05, BaseOMEAttrs, BaseZarrAttrs
+from ome_zarr_models.common.well_types import WellMeta
 
 __all__ = ["Well", "WellAttrs"]
 
 
-class WellAttrs(ome_zarr_models.v04.well.WellAttrs, BaseOMEAttrs):
+class WellAttrs(BaseOMEAttrs):
     """
     Attributes for a well.
     """
+
+    well: WellMeta
 
 
 class Well(GroupSpec[BaseZarrAttrs[WellAttrs], ArraySpec | GroupSpec], BaseGroupv05):  # type: ignore[misc]
