+++ src/ome_zarr_models/v04/well.py
@@ -6,20 +6,11 @@ from collections.abc import Generator
 
 from pydantic_zarr.v2 import ArraySpec, GroupSpec
 
-from ome_zarr_models.base import BaseAttrs
+from ome_zarr_models.common.well import WellAttrs
 from ome_zarr_models.v04.base import BaseGroupv04
 from ome_zarr_models.v04.image import Image
-from ome_zarr_models.v04.well_types import WellMeta
 
 __all__ = ["Well", "WellAttrs"]
-
-
-class WellAttrs(BaseAttrs):
-    """
-    Attributes for a well group.
-    """
-
-    well: WellMeta
 
 
 class Well(GroupSpec[WellAttrs, ArraySpec | GroupSpec], BaseGroupv04):  # type: ignore[misc]
